+++ src/ui/tasks_widget.rs
@@ -2,7 +2,8 @@
 
 use super::{
     AppBlockWidget, dialogs::DialogTrait, dialogs::ListDialog, draw_helper::DrawHelper, header::Header,
-    keyboard_handler::KeyboardHandler, mouse_handler::MouseHandler, shortcut::Shortcut, style, widgets::WidgetTrait,
+    keyboard_handler::KeyboardHandler, mouse_handler::MouseHandler, shortcut::Shortcut, style, widgets::TaskRow,
+    widgets::WidgetTrait,
 };
 use crate::{
     async_jobs::{AsyncJob, AsyncJobStorage},
@@ -13,7 +14,6 @@ use crate::{
     task::{self, Priority, State, Task as TaskTrait, datetime_to_str, due_group},
     task_patch::{DuePatchItem, PatchError, TaskPatch},
     types::ArcRwLock,
-    ui::widgets::{MarkdownLine, Text},
 };
 use async_trait::async_trait;
 use chrono::Local;
@@ -21,11 +21,10 @@ use crossterm::event::{KeyEvent, MouseEvent};
 use ratatui::{
     buffer::Buffer,
     layout::{Position, Rect, Size},
-    style::{Color, Style},
-    text::Text as RatatuiText,
+    text::Text,
     widgets::{Clear, ListState, Widget},
 };
-use std::{cmp::Ordering, slice::IterMut, sync::Arc};
+use std::{slice::IterMut, sync::Arc};
 use tokio::sync::{RwLock, broadcast};
 use tracing::{Instrument, Level};
 
@@ -57,141 +56,13 @@ pub trait TaskInfoViewerTrait: Send + Sync {
 
 type TaskInfoViewer = ArcRwLock<dyn TaskInfoViewerTrait>;
 
-struct TaskLineWidget {
-    task: Box<dyn TaskTrait>,
-    children: Vec<Box<dyn WidgetTrait>>,
-}
-
-impl TaskLineWidget {
-    pub fn new(t: &dyn TaskTrait, changed_tasks: &[TaskPatch]) -> Self {
-        let tz = Local::now().timezone();
-
-        let fg_color = {
-            match t.due() {
-                Some(d) => {
-                    let now = chrono::Utc::now().date_naive();
-                    match d.date_naive().cmp(&now) {
-                        Ordering::Less => style::OVERDUE_TASK_FG,
-                        Ordering::Equal => style::TODAY_TASK_FG,
-                        Ordering::Greater => style::FUTURE_TASK_FG,
-                    }
-                }
-                None => style::NO_DATE_TASK_FG,
-            }
-        };
-        let mut state = t.state();
-        let mut due = task::datetime_to_str(t.due(), &tz);
-        let mut priority = t.priority();
-        let mut uncommitted = false;
-        if let Some(patch) = changed_tasks.iter().find(|c| c.is_task(t)) {
-            uncommitted = !patch.is_empty();
-            if let Some(s) = &patch.state {
-                state = s.clone();
-            }
-            if let Some(d) = &patch.due {
-                due = d.to_string();
-            }
-            if let Some(p) = &patch.priority {
-                priority = p.clone();
-            }
-        }
-
-        let mut children: Vec<Box<dyn WidgetTrait>> = vec![
-            Box::new(Text::new(format!("[{state}] ").as_str())),
-            Box::new(MarkdownLine::new(t.text().as_str()).style(Style::default().fg(fg_color))),
-            Box::new(Text::new(format!(" (due: {due})").as_str()).style(Style::default().fg(Color::Blue))),
-            Box::new(
-                Text::new(format!(" (Priority: {priority})").as_str())
-                    .style(Style::default().fg(style::priority_color(&priority))),
-            ),
-            Box::new(Text::new(format!(" ({})", t.place()).as_str()).style(Style::default().fg(Color::Yellow))),
-        ];
-
-        if !t.description().unwrap_or_default().is_empty() {
-            children.push(Box::new(Text::new(" 💬")));
-        }
-
-        if uncommitted {
-            children.push(Box::new(Text::new(" 📤")));
-        }
-
-        Self {
-            task: t.clone_boxed(),
-            children,
-        }
-    }
-
-    fn task(&self) -> &dyn TaskTrait {
-        self.task.as_ref()
-    }
-}
-
-#[async_trait]
-impl WidgetTrait for TaskLineWidget {
-    async fn render(&mut self, area: Rect, buf: &mut Buffer) {
-        for child in self.children.iter_mut() {
-            child.render(area, buf).await;
-        }
-    }
-
-    fn size(&self) -> Size {
-        let mut result = Size::default();
-
-        for child in self.children.iter() {
-            result.width += child.size().width;
-            result.height = result.height.max(child.size().height);
-        }
-
-        result
-    }
-
-    fn set_style(&mut self, style: Style) {
-        for child in self.children.iter_mut() {
-            let mut s = child.style();
-            s.bg = None;
-            child.set_style(s.patch(style));
-        }
-    }
-
-    fn set_pos(&mut self, pos: Position) {
-        let mut x = pos.x;
-
-        for child in self.children.iter_mut() {
-            child.set_pos(Position::new(x, pos.y));
-            x += child.size().width;
-        }
-    }
-}
-
-#[async_trait]
-impl KeyboardHandler for TaskLineWidget {
-    async fn handle_key(&mut self, key: KeyEvent) -> bool {
-        for child in self.children.iter_mut() {
-            if child.handle_key(key).await {
-                return true;
-            }
-        }
-
-        false
-    }
-}
-
-#[async_trait]
-impl MouseHandler for TaskLineWidget {
-    async fn handle_mouse(&mut self, ev: &MouseEvent) {
-        for child in self.children.iter_mut() {
-            child.handle_mouse(ev).await;
-        }
-    }
-}
-
 pub struct TasksWidget {
     providers_storage: ArcRwLock<dyn ProvidersStorage<Provider>>,
     error_logger: ErrorLogger,
     task_info_viewer: TaskInfoViewer,
     all_tasks: Vec<Box<dyn TaskTrait>>,
     changed_tasks: Vec<TaskPatch>,
-    tasks: Vec<TaskLineWidget>,
+    tasks: Vec<TaskRow>,
     providers_filter: Vec<String>,
     projects_filter: Vec<String>,
     draw_helper: Option<DrawHelper>,
@@ -362,7 +233,7 @@ impl TasksWidget {
                 }
                 result
             })
-            .map(|t| TaskLineWidget::new(t.as_ref(), &self.changed_tasks))
+            .map(|t| TaskRow::new(t.as_ref(), &self.changed_tasks))
             .collect();
 
         self.state = if self.all_tasks.is_empty() {
@@ -750,7 +621,7 @@ impl WidgetTrait for TasksWidget {
 
         for (i, w) in self.tasks.iter_mut().enumerate() {
             if selected.is_some_and(|idx| idx == i) {
-                RatatuiText::from(">").render(
+                Text::from(">").render(
                     Rect {
                         x: area.x,
                         y,
