+++ src/ui/widgets.rs
@@ -3,11 +3,13 @@
 pub mod hyperlink_widget;
 pub mod line_edit;
 pub mod markdown_line;
+pub mod task_row;
 pub mod text;
 pub mod widget;
 
 pub use hyperlink_widget::HyperlinkWidget;
 pub use line_edit::LineEdit;
 pub use markdown_line::MarkdownLine;
+pub use task_row::TaskRow;
 pub use text::Text;
 pub use widget::WidgetTrait;
